+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/component/CertiEmptySection.kt
