+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/component/ResumeTextField.kt
@@ -5,48 +5,22 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.sopt.certi.R
 import org.sopt.certi.core.util.screenHeightDp
-import org.sopt.certi.ui.theme.CERTITheme
 import org.sopt.certi.ui.theme.CertiTheme
 
 @Composable
-fun ResumeInputSection(
-    title: String,
-    value: String,
-    onValueChange: (String) -> Unit,
-    maxLength: Int,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier.fillMaxWidth()
-    ) {
-        ResumeInputTitle(title)
-        Spacer(modifier = Modifier.height(screenHeightDp(24.dp)))
-        ResumeTextField(
-            value = value,
-            onValueChange = onValueChange,
-            maxLength = maxLength
-        )
-    }
-}
-
-@Composable
-private fun ResumeTextField(
+fun ResumeTextField(
     value: String,
     onValueChange: (String) -> Unit,
     maxLength: Int,
@@ -94,18 +68,4 @@ private fun ResumeTextField(
             }
         }
     )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun ResumeInputSectionPreview() {
-    val text = remember { mutableStateOf("") }
-    CERTITheme {
-        ResumeInputSection(
-            title = "근무회사",
-            value = text.value,
-            onValueChange = { text.value = it },
-            maxLength = 10
-        )
-    }
 }
